Get triangle animation working!
Using what was in repulsion (with some teasing apart and comments to help myself), I figured out how to move the triangles and the dots at the same time. This commit creates objects for them that have similar methods, though the triangle's are necessarily more complicated. The friction calculation is slightly tweaked, and we also check for triangles with tiny angles so we can redraw them until they look nicer. It's cool I'm proud of it! Could probably do a little more cleanup and DRY-ing up, but certainly it's a good start.